Create Injectable GovUK and Main Templates for individual and migrate Home view
