+++ app/views/Home.scala.html
@@ -17,7 +17,11 @@
 @import java.time.LocalDate
 @import config.{FrontendAppConfig, ItvcErrorHandler, ItvcHeaderCarrierForPartialsConverter}
 @import implicits.ImplicitDateFormatter
-@import views.html.templates.main_template
+@import views.html.templates.MainTemplate
+
+@this(
+    mainTemplate: MainTemplate
+)
 
 @(nextPaymentDueDate: Option[LocalDate], nextUpdate: LocalDate, overDuePayments: Option[Int], overDueUpdates: Option[Int], paymentHistoryEnabled: Boolean, ITSASubmissionIntegrationEnabled: Boolean,implicitDateFormatter: ImplicitDateFormatter)(implicit request: Request[_], messages: Messages, appConfig: config.FrontendAppConfig, user: auth.MtdItUser[_])
 @import implicitDateFormatter.longDate
@@ -103,9 +107,9 @@
     }
 }
 
-@main_template(
-title = messages("home.heading"),
-mainClass = Some("full-width")
+@mainTemplate(
+    title = messages("home.heading"),
+    mainClass = Some("full-width")
 ) {
     <h1 class="heading-xlarge">
         <span id="income-tax-heading">@messages("home.heading")</span>
