+++ test/controllers/agent/IncomeSummaryControllerSpec.scala
@@ -23,7 +23,7 @@ import implicits.ImplicitDateFormatterImpl
 import mocks.MockItvcErrorHandler
 import mocks.auth.MockFrontendAuthorisedFunctions
 import mocks.services.{MockCalculationService, MockIncomeSourceDetailsService}
-import mocks.views.MockIncomeSummary
+import mocks.views.agent.MockIncomeSummary
 import models.calculation.CalcDisplayError
 import play.api.http.Status
 import play.api.mvc.MessagesControllerComponents
