+++ test/mocks/views/agent/MockUTRError.scala
@@ -14,30 +14,27 @@
  * limitations under the License.
  */
 
-package mocks.views
+package mocks.views.agent
 
-import java.time.LocalDate
-
-import org.mockito.ArgumentMatchers.{any, eq => matches}
+import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.{reset, when}
 import org.scalatest.{BeforeAndAfterEach, Suite}
 import org.scalatestplus.mockito.MockitoSugar
 import play.twirl.api.Html
-import views.html.agent.Home
+import views.html.agent.UTRError
 
-trait MockHome extends BeforeAndAfterEach with MockitoSugar {
+trait MockUTRError extends BeforeAndAfterEach with MockitoSugar {
   self: Suite =>
 
-  val home: Home = mock[Home]
+  val utrError: UTRError = mock[UTRError]
 
   override def beforeEach(): Unit = {
     super.beforeEach()
-    reset(home)
+    reset(utrError)
   }
 
-  def mockHome(nextPaymentOrOverdue: Option[Either[(LocalDate, Boolean), Int]],
-                          nextUpdateOrOverdue: Either[(LocalDate, Boolean), Int])(response: Html): Unit = {
-    when(home.apply(matches(nextPaymentOrOverdue), matches(nextUpdateOrOverdue), any(), any(), any(), any())(any(), any(), any(), any()))
+  def mockUTRError(response: Html): Unit = {
+    when(utrError.apply(any(), any())(any(), any(), any()))
       .thenReturn(response)
   }
