+++ test/views/HomePageViewSpec.scala
@@ -29,6 +29,7 @@ import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import play.twirl.api.HtmlFormat
 import testUtils.TestSupport
+import views.html.Home
 
 import java.time.LocalDate
 
@@ -59,7 +60,8 @@ class HomePageViewSpec extends TestSupport with FeatureSwitching {
   class Setup(paymentDueDate: Option[LocalDate] = Some(nextPaymentDueDate), overDuePayments: Option[Int] = Some(0),
               overDueUpdates: Option[Int] = Some(0), paymentHistoryEnabled: Boolean = true, ITSASubmissionIntegrationEnabled: Boolean = true) {
 
-    lazy val page: HtmlFormat.Appendable = views.html.home(
+    val home: Home = app.injector.instanceOf[Home]
+    lazy val page: HtmlFormat.Appendable = home(
       nextPaymentDueDate = paymentDueDate,
       nextUpdate = updateDate,
       overDuePayments = overDuePayments,
